Intern_Submissions/Kanaka_Durga/autonomous-cognitive-engine/requirements.txt
streamlit==1.32.2
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
